server/requirements.txt
openenv-core>=0.2.3
fastapi
uvicorn
pydantic
openai
python-dotenv



